Use eslint:recommended and Prettier defaults (#132)
Long time ago this project adopted the "standard" style because I wanted an opinionated tool that handled all the linting and styling and didn't want to maintain the rules myself. Now that Prettier exists, I'm opting to just rely on Prettier for styling and use eslint:recommended checks for linting.

This allows getting rid of several additional packages, which is useful! Fewer packages, the less maintenance for keeping things up-to-date.

Unfortunately (or fortunately) using defaults involves changes to every JS file in the project.

It is worth it in my opinion however. I'd like to support the defaults, and if anything my preference leans semicolon.

With Postgrator 5, I'm anticipating gutting and simplifying a lot of the package anyways, so its a good time to do it.